+++ src/components/Menu/Menu.js
@@ -1,35 +1,15 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import ControlledMenu from './ControlledMenu';
-import cx from 'classnames';
-import enhanceWithClickOutside from 'react-click-outside';
-import omit from 'lodash/omit';
-import { Button, IconButton } from '../../';
 
 // https://github.com/oliviertassinari/babel-plugin-transform-react-remove-prop-types#is-it-safe
 const propTypes = {
   /**
    * The button that triggers the dropdown menu
-   * ```
-   * {
-   *    icon: 'settings',
-   *    sprite: 'utility',
-   *    title: 'Click me',
-   *    noBorder: true,
-   * }
-   * ```
    */
-  button: PropTypes.shape({
-    brand: PropTypes.bool,
-    icon: PropTypes.string.isRequired,
-    neutral: PropTypes.bool,
-    noBorder: PropTypes.bool,
-    sprite: PropTypes.string.isRequired,
-    title: PropTypes.string,
-    tooltip: PropTypes.string,
-  }),
+  button: PropTypes.element.isRequired,
   /**
-   * one MenuList or many of them
+   * should be MenuItems or MenuSubHeaders
    */
   children: PropTypes.node.isRequired,
   /**
@@ -37,18 +17,9 @@ const propTypes = {
    */
   className: PropTypes.string,
   /**
-   * fully customizable dropdown trigger button, use this instead of the button
-   * shape if needed
+   * menu should start open
    */
-  customButton: PropTypes.element,
-  /**
-   * adds disabled attribute to menu button
-   */
-  disabled: PropTypes.bool,
-  /**
-   * forces open or closed state, is needed when using a custom button
-   */
-  isOpen: PropTypes.bool,
+  defaultOpen: PropTypes.bool,
   /**
    * indicates that this is the last element inside a button group and renders
    * the required css class
@@ -75,8 +46,7 @@ class Menu extends Component {
     button: null,
     className: null,
     customButton: null,
-    disabled: false,
-    isOpen: false,
+    defaultOpen: false,
     last: false,
     nubbin: false,
     position: 'top-left',
@@ -85,7 +55,12 @@ class Menu extends Component {
 
   constructor(props, { cssPrefix }) {
     super(props, { cssPrefix });
-    this.state = { open: this.props.isOpen };
+    this.state = { open: this.props.defaultOpen };
+  }
+
+  getButton = () => {
+    const { button } = this.props;
+    return React.cloneElement(button, { onClick: this.toggle });
   }
 
   toggle = () => {
@@ -96,55 +71,15 @@ class Menu extends Component {
     this.setState({ open: false });
   }
 
-  button = () => {
-    const { button, customButton, disabled } = this.props;
-
-    if (customButton) return customButton;
-
-    let buttonFlavor = null;
-
-    if (button.brand) { buttonFlavor = 'brand'; }
-    if (button.neutral) { buttonFlavor = 'neutral'; }
-
-    if (!button.title) {
-      return (
-        <IconButton
-          aria-haspopup="true"
-          icon={button.icon}
-          disabled={disabled}
-          flavor={buttonFlavor}
-          onClick={this.toggle}
-          sprite={button.sprite}
-          border={button.noBorder ? null : 'filled'}
-          container={button.noBorder}
-          title={button.tooltip}
-        />
-      );
-    }
-
-    return (
-      <Button
-        aria-haspopup="true"
-        flavor={buttonFlavor}
-        icon={button.icon}
-        iconPosition="right"
-        sprite={button.sprite}
-        onClick={this.toggle}
-        title={button.title}
-      />
-    );
-  }
-
   render() {
-    const { isOpen, ...rest } = this.props;
-
     const { open } = this.state;
 
     return (
       <ControlledMenu
-        isOpen={isOpen || open}
+        {...this.props}
+        button={this.getButton()}
+        isOpen={open}
         onMenuClick={this.toggle}
-        {...rest}
       />
     );
   }